+++ lw6_t3/src/graphics_lib/CanvasInterface.php
@@ -5,6 +5,6 @@ namespace GraphicsLib;
 interface CanvasInterface
 {
     public function setColor(string $color): void;
-    public function moveTo(int $x, int $y): void;
-    public function lineTo(int $x, int $y): void;
+    public function moveTo(float $x, float $y): void;
+    public function lineTo(float $x, float $y): void;
 }
